Refs #33078 - extract SelectAllCheckbox into its own component
